// src/client/types.gen.ts
export type IdentityRoleDto = {
	readonly extraProperties?: {
		[key: string]: unknown;
	} | null;
	id?: string;
	name?: string | null;
	isDefault?: boolean;
	isStatic?: boolean;
	isPublic?: boolean;
	concurrencyStamp?: string | null;
	creationTime?: string;
};